# api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/v1/achievements": {
            "get": {
                "description": "Retrieve list of courses.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get list of courses.",
                "operationId": "get-courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Course"
                            }
                        }
                    },
                    "400": {}
                }
            }
        },
        "/api/v1/ping": {
            "get": {
                "description": "ping",
                "summary": "Ping",
                "responses": {
                    "200": {},
                    "400": {}
                }
            }
        },
        "/api/v1/tasks/{user_id}": {
            "get": {
                "description": "Retrieve list of tasks using user identifier.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get list of tasks by user identifier.",
                "operationId": "get-tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Task"
                            }
                        }
                    },
                    "400": {}
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "Retrieve a user info.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get user.",
                "operationId": "get-user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {}
                }
            }
        }
    },
    "definitions": {
        "model.Achievement": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "expirationDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Course": {
            "type": "object",
            "properties": {
                "award": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.Task": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "expirationDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "managerID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "taskPriority": {
                    "type": "object",
                    "$ref": "#/definitions/model.TaskPriority"
                },
                "taskPriorityID": {
                    "type": "integer"
                },
                "taskStatus": {
                    "type": "object",
                    "$ref": "#/definitions/model.TaskStatus"
                },
                "taskStatusID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.TaskPriority": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.TaskStatus": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "skill": {
                    "type": "object",
                    "$ref": "#/definitions/postgres.Jsonb"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Task"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.Jsonb": {
            "type": "object"
        }
    }
}